+++ tasks/levonychev_i_min_val_rows_matrix/seq/src/ops_seq.cpp
@@ -16,31 +16,31 @@ LevonychevIMinValRowsMatrixSEQ::LevonychevIMinValRowsMatrixSEQ(const InType &in)
 
 bool LevonychevIMinValRowsMatrixSEQ::ValidationImpl() {
   const size_t vector_size_ = std::get<0>(GetInput()).size();
-  const size_t ROWS = std::get<1>(GetInput());
-  const size_t COLS = std::get<2>(GetInput());
+  const int ROWS = std::get<1>(GetInput());
+  const int COLS = std::get<2>(GetInput());
   if (vector_size_ == 0 || ROWS == 0 || COLS == 0) {
     return false;
   }
-  if (vector_size_ != ROWS * COLS) {
+  if (vector_size_ != static_cast<size_t>(ROWS * COLS)) {
     return false;
   }
   return true;
 }
 
 bool LevonychevIMinValRowsMatrixSEQ::PreProcessingImpl() {
-  GetOutput().resize(std::get<1>(GetInput()));
+  GetOutput().resize(static_cast<size_t>(std::get<1>(GetInput())));
   return true;
 }
 
 bool LevonychevIMinValRowsMatrixSEQ::RunImpl() {
   const std::vector<int> &matrix = std::get<0>(GetInput());
-  const size_t ROWS = std::get<1>(GetInput());
-  const size_t COLS = std::get<2>(GetInput());
+  const int ROWS = std::get<1>(GetInput());
+  const int COLS = std::get<2>(GetInput());
   OutType& result = GetOutput();
   
-  for (size_t i = 0; i < ROWS; ++i) {
+  for (int i = 0; i < ROWS; ++i) {
     int min_val = matrix[COLS * i];
-    for (size_t j = 1; j < COLS; ++j) {
+    for (int j = 1; j < COLS; ++j) {
       if (matrix[COLS * i + j] < min_val) {
         min_val = matrix[COLS * i + j];
       }
